Move more embedded boundary methods out of WarpX class and rename EmbeddedBoundary.H/cpp
