+++ Source/EmbeddedBoundary/CMakeLists.txt
@@ -2,7 +2,7 @@ foreach(D IN LISTS WarpX_DIMS)
     warpx_set_suffix_dims(SD ${D})
     target_sources(lib_${SD}
       PRIVATE
-        EmbeddedBoundary.cpp
+        EmbeddedBoundaryInit.cpp
         Enabled.cpp
         WarpXInitEB.cpp
         WarpXFaceExtensions.cpp
