--- Source/EmbeddedBoundary/EmbeddedBoundary.cpp
@@ -1,200 +1,0 @@
-/* Copyright 2021-2025 Lorenzo Giacomel, Luca Fedeli
- *
- * This file is part of WarpX.
- *
- * License: BSD-3-Clause-LBNL
- */
-
-#include "Enabled.H"
-
-#ifdef AMREX_USE_EB
-
-#include "EmbeddedBoundary.H"
-
-#include "Utils/TextMsg.H"
-
-#include <AMReX_BLProfiler.H>
-#include <AMReX_Box.H>
-#include <AMReX_GpuLaunch.H>
-#include <AMReX_GpuQualifiers.H>
-#include <AMReX_MFIter.H>
-#include <AMReX_MultiCutFab.H>
-
-namespace web = warpx::embedded_boundary;
-
-void
-web::ComputeEdgeLengths (
-    ablastr::fields::VectorField& edge_lengths,
-    const amrex::EBFArrayBoxFactory& eb_fact)
-{
-    BL_PROFILE("ComputeEdgeLengths");
-
-#if !defined(WARPX_DIM_3D) && !defined(WARPX_DIM_XZ) && !defined(WARPX_DIM_RZ)
-    WARPX_ABORT_WITH_MESSAGE("ComputeEdgeLengths only implemented in 2D and 3D");
-#endif
-
-    auto const &flags = eb_fact.getMultiEBCellFlagFab();
-    auto const &edge_centroid = eb_fact.getEdgeCent();
-    for (int idim = 0; idim < 3; ++idim){
-#if defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-        if (idim == 1) {
-            edge_lengths[1]->setVal(0.);
-            continue;
-        }
-#endif
-        for (amrex::MFIter mfi(flags); mfi.isValid(); ++mfi){
-            amrex::Box const box = mfi.tilebox(edge_lengths[idim]->ixType().toIntVect(),
-                                               edge_lengths[idim]->nGrowVect());
-            amrex::FabType const fab_type = flags[mfi].getType(box);
-            auto const &edge_lengths_dim = edge_lengths[idim]->array(mfi);
-
-            if (fab_type == amrex::FabType::regular) {
-                // every cell in box is all regular
-                amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                    edge_lengths_dim(i, j, k) = 1.;
-                });
-            } else if (fab_type == amrex::FabType::covered) {
-                // every cell in box is all covered
-                amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                    edge_lengths_dim(i, j, k) = 0.;
-                });
-            } else {
-#if defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-                int idim_amrex = idim;
-                if (idim == 2) { idim_amrex = 1; }
-                auto const &edge_cent = edge_centroid[idim_amrex]->const_array(mfi);
-#elif defined(WARPX_DIM_3D)
-                auto const &edge_cent = edge_centroid[idim]->const_array(mfi);
-#endif
-                amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                    if (edge_cent(i, j, k) == amrex::Real(-1.0)) {
-                        // This edge is all covered
-                        edge_lengths_dim(i, j, k) = 0.;
-                    } else if (edge_cent(i, j, k) == amrex::Real(1.0)) {
-                        // This edge is all open
-                        edge_lengths_dim(i, j, k) = 1.;
-                    } else {
-                        // This edge is cut.
-                        edge_lengths_dim(i, j, k) = 1 - amrex::Math::abs(amrex::Real(2.0)
-                                                                        * edge_cent(i, j, k));
-                    }
-
-                });
-            }
-        }
-    }
-}
-
-
-void
-web::ComputeFaceAreas (
-    ablastr::fields::VectorField& face_areas,
-    const amrex::EBFArrayBoxFactory& eb_fact)
-{
-    BL_PROFILE("ComputeFaceAreas");
-
-#if !defined(WARPX_DIM_3D) && !defined(WARPX_DIM_XZ) && !defined(WARPX_DIM_RZ)
-    WARPX_ABORT_WITH_MESSAGE("ComputeFaceAreas only implemented in 2D and 3D");
-#endif
-
-    auto const &flags = eb_fact.getMultiEBCellFlagFab();
-#if defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-    //In 2D the volume frac is actually the area frac.
-    auto const &area_frac = eb_fact.getVolFrac();
-#elif defined(WARPX_DIM_3D)
-    auto const &area_frac = eb_fact.getAreaFrac();
-#endif
-
-    for (int idim = 0; idim < 3; ++idim) {
-#if defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-        if (idim == 0 || idim == 2) {
-            face_areas[idim]->setVal(0.);
-            continue;
-        }
-#endif
-        for (amrex::MFIter mfi(flags); mfi.isValid(); ++mfi) {
-            amrex::Box const box = mfi.tilebox(face_areas[idim]->ixType().toIntVect(),
-                                               face_areas[idim]->nGrowVect());
-            amrex::FabType const fab_type = flags[mfi].getType(box);
-            auto const &face_areas_dim = face_areas[idim]->array(mfi);
-            if (fab_type == amrex::FabType::regular) {
-                // every cell in box is all regular
-                amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                    face_areas_dim(i, j, k) = amrex::Real(1.);
-                });
-            } else if (fab_type == amrex::FabType::covered) {
-                // every cell in box is all covered
-                amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                    face_areas_dim(i, j, k) = amrex::Real(0.);
-                });
-            } else {
-#if defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-                auto const &face = area_frac.const_array(mfi);
-#elif defined(WARPX_DIM_3D)
-                auto const &face = area_frac[idim]->const_array(mfi);
-#endif
-                amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                    face_areas_dim(i, j, k) = face(i, j, k);
-                });
-            }
-        }
-    }
-}
-
-void
-web::ScaleEdges (
-    ablastr::fields::VectorField& edge_lengths,
-    const std::array<amrex::Real,3>& cell_size)
-{
-    BL_PROFILE("ScaleEdges");
-
-#if !defined(WARPX_DIM_3D) && !defined(WARPX_DIM_XZ) && !defined(WARPX_DIM_RZ)
-    WARPX_ABORT_WITH_MESSAGE("ScaleEdges only implemented in 2D and 3D");
-#endif
-
-    for (int idim = 0; idim < 3; ++idim){
-#if defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-        if (idim == 1) { continue; }
-#endif
-        for (amrex::MFIter mfi(*edge_lengths[0]); mfi.isValid(); ++mfi) {
-            const amrex::Box& box = mfi.tilebox(edge_lengths[idim]->ixType().toIntVect(),
-                                                edge_lengths[idim]->nGrowVect() );
-            auto const &edge_lengths_dim = edge_lengths[idim]->array(mfi);
-            amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                edge_lengths_dim(i, j, k) *= cell_size[idim];
-            });
-        }
-    }
-}
-
-
-void
-web::ScaleAreas (
-    ablastr::fields::VectorField& face_areas,
-    const std::array<amrex::Real,3>& cell_size)
-{
-    BL_PROFILE("ScaleAreas");
-
-#if !defined(WARPX_DIM_3D) && !defined(WARPX_DIM_XZ) && !defined(WARPX_DIM_RZ)
-    WARPX_ABORT_WITH_MESSAGE("ScaleAreas only implemented in 2D and 3D");
-#endif
-
-    for (int idim = 0; idim < 3; ++idim) {
-#if defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-        if (idim == 0 || idim == 2) { continue; }
-#endif
-        for (amrex::MFIter mfi(*face_areas[0]); mfi.isValid(); ++mfi) {
-            const amrex::Box& box = mfi.tilebox(face_areas[idim]->ixType().toIntVect(),
-                                                face_areas[idim]->nGrowVect() );
-            amrex::Real const full_area = cell_size[(idim+1)%3]*cell_size[(idim+2)%3];
-            auto const &face_areas_dim = face_areas[idim]->array(mfi);
-
-            amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                face_areas_dim(i, j, k) *= full_area;
-            });
-
-        }
-    }
-}
-
-#endif
