+++ Source/EmbeddedBoundary/WarpXInitEB.cpp
@@ -126,312 +126,6 @@ WarpX::InitEB ()
 #ifdef AMREX_USE_EB
 
 void
-WarpX::MarkReducedShapeCells (
-    std::unique_ptr<amrex::iMultiFab> & eb_reduce_particle_shape,
-    amrex::EBFArrayBoxFactory const & eb_fact,
-    int const particle_shape_order )
-{
-    // Pre-fill array with 0, including in the ghost cells outside of the domain.
-    // (The guard cells in the domain will be updated by `FillBoundary` at the end of this function.)
-    eb_reduce_particle_shape->setVal(0, eb_reduce_particle_shape->nGrow());
-
-    // Extract structures for embedded boundaries
-    amrex::FabArray<amrex::EBCellFlagFab> const& eb_flag = eb_fact.getMultiEBCellFlagFab();
-
-#ifdef AMREX_USE_OMP
-#pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
-#endif
-    for (amrex::MFIter mfi(*eb_reduce_particle_shape); mfi.isValid(); ++mfi) {
-
-        const amrex::Box& box = mfi.tilebox();
-        amrex::Array4<int> const & eb_reduce_particle_shape_arr = eb_reduce_particle_shape->array(mfi);
-
-        // Check if the box (including one layer of guard cells) contains a mix of covered and regular cells
-        const amrex::Box& eb_info_box = mfi.tilebox(amrex::IntVect::TheCellVector()).grow(1);
-        amrex::FabType const fab_type = eb_flag[mfi].getType( eb_info_box );
-
-        if (fab_type == amrex::FabType::regular) { // All cells in the box are regular
-
-            // Every cell in box is regular: do not reduce particle shape in any cell
-            amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                eb_reduce_particle_shape_arr(i, j, k) = 0;
-            });
-
-        } else if (fab_type == amrex::FabType::covered) { // All cells in the box are covered
-
-            // Every cell in box is fully covered: reduce particle shape
-            amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                eb_reduce_particle_shape_arr(i, j, k) = 1;
-            });
-
-        } else { // The box contains a mix of covered and regular cells
-
-            auto const & flag = eb_flag[mfi].array();
-
-            amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-
-                // Check if any of the neighboring cells over which the particle shape might extend
-                // are either partially or fully covered. In this case, set eb_reduce_particle_shape_arr
-                // to one for this cell, to indicate that the particle should use an order 1 shape
-                // (This ensures that the particle never deposits any charge in a partially or
-                // fully covered cell, even with higher-order shapes)
-                // Note: in the code below `particle_shape_order/2` corresponds to the number of neighboring cells
-                // over which the shape factor could extend, in each direction.
-                int const i_start = i-particle_shape_order/2;
-                int const i_end = i+particle_shape_order/2;
-#if AMREX_SPACEDIM > 1
-                int const j_start = j-particle_shape_order/2;
-                int const j_end = j+particle_shape_order/2;
-#else
-                int const j_start = j;
-                int const j_end = j;
-#endif
-#if AMREX_SPACEDIM > 2
-                int const k_start = k-particle_shape_order/2;
-                int const k_end = k+particle_shape_order/2;
-#else
-                int const k_start = k;
-                int const k_end = k;
-#endif
-                int reduce_shape = 0;
-                for (int i_cell = i_start; i_cell <= i_end; ++i_cell) {
-                    for (int j_cell = j_start; j_cell <= j_end; ++j_cell) {
-                        for (int k_cell = k_start; k_cell <= k_end; ++k_cell) {
-                            // `isRegular` returns `false` if the cell is either partially or fully covered.
-                            if ( !flag(i_cell, j_cell, k_cell).isRegular() ) {
-                                reduce_shape = 1;
-                            }
-                        }
-                    }
-                }
-                eb_reduce_particle_shape_arr(i, j, k) = reduce_shape;
-            });
-
-        }
-
-    }
-
-    // FillBoundary to set the values in the guard cells
-    eb_reduce_particle_shape->FillBoundary(Geom(0).periodicity());
-
-}
-
-void
-WarpX::MarkUpdateCellsStairCase (
-    std::array< std::unique_ptr<amrex::iMultiFab>,3> & eb_update,
-    ablastr::fields::VectorField const& field,
-    amrex::EBFArrayBoxFactory const & eb_fact )
-{
-
-    using ablastr::fields::Direction;
-    using warpx::fields::FieldType;
-
-    // Extract structures for embedded boundaries
-    amrex::FabArray<amrex::EBCellFlagFab> const& eb_flag = eb_fact.getMultiEBCellFlagFab();
-
-    for (int idim = 0; idim < 3; ++idim) {
-
-#ifdef AMREX_USE_OMP
-#pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
-#endif
-        for (amrex::MFIter mfi(*field[idim]); mfi.isValid(); ++mfi) {
-
-            const amrex::Box& box = mfi.tilebox();
-            amrex::Array4<int> const & eb_update_arr = eb_update[idim]->array(mfi);
-
-            // Check if the box (including one layer of guard cells) contains a mix of covered and regular cells
-            const amrex::Box& eb_info_box = mfi.tilebox(amrex::IntVect::TheCellVector()).grow(1);
-            amrex::FabType const fab_type = eb_flag[mfi].getType( eb_info_box );
-
-            if (fab_type == amrex::FabType::regular) { // All cells in the box are regular
-
-                // Every cell in box is regular: update field in every cell
-                amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                    eb_update_arr(i, j, k) = 1;
-                });
-
-            } else if (fab_type == amrex::FabType::covered) { // All cells in the box are covered
-
-                // Every cell in box is fully covered: do not update field
-                amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                    eb_update_arr(i, j, k) = 0;
-                });
-
-            } else { // The box contains a mix of covered and regular cells
-
-                auto const & flag = eb_flag[mfi].array();
-                auto index_type = field[idim]->ixType();
-
-                amrex::ParallelFor(box, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-
-                    // Stair-case approximation: If neighboring cells of this gridpoint
-                    // are either partially or fully covered: do not update field
-
-                    // The number of cells that we need to check depend on the index type
-                    // of the `eb_update_arr` in each direction.
-                    // If `eb_update_arr` is nodal in a given direction, we need to check the cells
-                    // to the left and right of this nodal gridpoint.
-                    // For instance, if `eb_update_arr` is nodal in the first dimension, we need
-                    // to check the cells at index i-1 and at index i, since, with AMReX indexing conventions,
-                    // these are the neighboring cells for the nodal gripoint at index i.
-                    // If `eb_update_arr` is cell-centerd in a given direction, we only need to check
-                    // the cell at the same position (e.g., in the first dimension: the cell at index i).
-                    int const i_start = ( index_type.nodeCentered(0) )? i-1 : i;
-#if AMREX_SPACEDIM > 1
-                    int const j_start = ( index_type.nodeCentered(1) )? j-1 : j;
-#else
-                    int const j_start = j;
-#endif
-#if AMREX_SPACEDIM > 2
-                    int const k_start = ( index_type.nodeCentered(2) )? k-1 : k;
-#else
-                    int const k_start = k;
-#endif
-                    // Loop over neighboring cells
-                    int eb_update_flag = 1;
-                    for (int i_cell = i_start; i_cell <= i; ++i_cell) {
-                        for (int j_cell = j_start; j_cell <= j; ++j_cell) {
-                            for (int k_cell = k_start; k_cell <= k; ++k_cell) {
-                                // If one of the neighboring is either partially or fully covered
-                                // (i.e. if they are not regular cells), do not update field
-                                // (`isRegular` returns `false` if the cell is either partially or fully covered.)
-                                if ( !flag(i_cell, j_cell, k_cell).isRegular() ) {
-                                    eb_update_flag = 0;
-                                }
-                            }
-                        }
-                    }
-                    eb_update_arr(i, j, k) = eb_update_flag;
-                });
-
-            }
-
-        }
-
-    }
-
-}
-
-void
-WarpX::MarkUpdateECellsECT (
-    std::array< std::unique_ptr<amrex::iMultiFab>,3> & eb_update_E,
-    ablastr::fields::VectorField const& edge_lengths )
-{
-
-#ifdef AMREX_USE_OMP
-#pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
-#endif
-    for ( amrex::MFIter mfi(*eb_update_E[0], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
-
-        const amrex::Box& tbx = mfi.tilebox( eb_update_E[0]->ixType().toIntVect(), eb_update_E[0]->nGrowVect() );
-        const amrex::Box& tby = mfi.tilebox( eb_update_E[1]->ixType().toIntVect(), eb_update_E[1]->nGrowVect() );
-        const amrex::Box& tbz = mfi.tilebox( eb_update_E[2]->ixType().toIntVect(), eb_update_E[2]->nGrowVect() );
-
-        amrex::Array4<int> const & eb_update_Ex_arr = eb_update_E[0]->array(mfi);
-        amrex::Array4<int> const & eb_update_Ey_arr = eb_update_E[1]->array(mfi);
-        amrex::Array4<int> const & eb_update_Ez_arr = eb_update_E[2]->array(mfi);
-
-        amrex::Array4<amrex::Real> const & lx_arr = edge_lengths[0]->array(mfi);
-        amrex::Array4<amrex::Real> const & lz_arr = edge_lengths[2]->array(mfi);
-#if defined(WARPX_DIM_3D)
-        amrex::Array4<amrex::Real> const & ly_arr = edge_lengths[1]->array(mfi);
-#elif defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-        amrex::Dim3 const lx_lo = amrex::lbound(lx_arr);
-        amrex::Dim3 const lx_hi = amrex::ubound(lx_arr);
-        amrex::Dim3 const lz_lo = amrex::lbound(lz_arr);
-        amrex::Dim3 const lz_hi = amrex::ubound(lz_arr);
-#endif
-
-        amrex::ParallelFor (tbx, tby, tbz,
-            [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                // Do not update Ex if the edge on which it lives is fully covered
-                eb_update_Ex_arr(i, j, k) = (lx_arr(i, j, k) == 0)? 0 : 1;
-            },
-            [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-#ifdef WARPX_DIM_3D
-                // In 3D: Do not update Ey if the edge on which it lives is fully covered
-                eb_update_Ey_arr(i, j, k) = (ly_arr(i, j, k) == 0)? 0 : 1;
-#elif defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-                // In XZ and RZ: Ey is associated with a mesh node,
-                // so we need to check if  the mesh node is covered
-                if((lx_arr(std::min(i  , lx_hi.x), std::min(j  , lx_hi.y), k)==0)
-                 ||(lx_arr(std::max(i-1, lx_lo.x), std::min(j  , lx_hi.y), k)==0)
-                 ||(lz_arr(std::min(i  , lz_hi.x), std::min(j  , lz_hi.y), k)==0)
-                 ||(lz_arr(std::min(i  , lz_hi.x), std::max(j-1, lz_lo.y), k)==0)) {
-                    eb_update_Ey_arr(i, j, k) = 0;
-                } else {
-                    eb_update_Ey_arr(i, j, k) = 1;
-                }
-#endif
-            },
-            [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                // Do not update Ez if the edge on which it lives is fully covered
-                eb_update_Ez_arr(i, j, k) = (lz_arr(i, j, k) == 0)? 0 : 1;
-            }
-        );
-
-    }
-}
-
-void
-WarpX::MarkUpdateBCellsECT (
-    std::array< std::unique_ptr<amrex::iMultiFab>,3> & eb_update_B,
-    ablastr::fields::VectorField const& face_areas,
-    ablastr::fields::VectorField const& edge_lengths )
-{
-
-#ifdef AMREX_USE_OMP
-#pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
-#endif
-    for ( amrex::MFIter mfi(*eb_update_B[0], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
-
-        const amrex::Box& tbx = mfi.tilebox( eb_update_B[0]->ixType().toIntVect(), eb_update_B[0]->nGrowVect() );
-        const amrex::Box& tby = mfi.tilebox( eb_update_B[1]->ixType().toIntVect(), eb_update_B[1]->nGrowVect() );
-        const amrex::Box& tbz = mfi.tilebox( eb_update_B[2]->ixType().toIntVect(), eb_update_B[2]->nGrowVect() );
-
-        amrex::Array4<int> const & eb_update_Bx_arr = eb_update_B[0]->array(mfi);
-        amrex::Array4<int> const & eb_update_By_arr = eb_update_B[1]->array(mfi);
-        amrex::Array4<int> const & eb_update_Bz_arr = eb_update_B[2]->array(mfi);
-
-#ifdef WARPX_DIM_3D
-        amrex::Array4<amrex::Real> const & Sx_arr = face_areas[0]->array(mfi);
-        amrex::Array4<amrex::Real> const & Sy_arr = face_areas[1]->array(mfi);
-        amrex::Array4<amrex::Real> const & Sz_arr = face_areas[2]->array(mfi);
-        amrex::ignore_unused(edge_lengths);
-#elif defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-        amrex::Array4<amrex::Real> const & Sy_arr = face_areas[1]->array(mfi);
-        amrex::Array4<amrex::Real> const & lx_arr = edge_lengths[0]->array(mfi);
-        amrex::Array4<amrex::Real> const & lz_arr = edge_lengths[2]->array(mfi);
-#endif
-        amrex::ParallelFor (tbx, tby, tbz,
-            [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-#ifdef WARPX_DIM_3D
-                // In 3D: do not update Bx if the face on which it lives is fully covered
-                eb_update_Bx_arr(i, j, k) = (Sx_arr(i, j, k) == 0)? 0 : 1;
-#elif defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-                //In XZ and RZ, Bx lives on a z-edge ; do not update if fully covered
-                eb_update_Bx_arr(i, j, k) = (lz_arr(i, j, k) == 0)? 0 : 1;
-#endif
-            },
-            [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-                // Do not update By if the face on which it lives is fully covered
-                eb_update_By_arr(i, j, k) = (Sy_arr(i, j, k) == 0)? 0 : 1;
-            },
-            [=] AMREX_GPU_DEVICE (int i, int j, int k) {
-#ifdef WARPX_DIM_3D
-                // In 3D: do not update Bz if the face on which it lives is fully covered
-                eb_update_Bz_arr(i, j, k) = (Sz_arr(i, j, k) == 0)? 0 : 1;
-#elif defined(WARPX_DIM_XZ) || defined(WARPX_DIM_RZ)
-                //In XZ and RZ, Bz lives on a x-edge ; do not update if fully covered
-                eb_update_Bz_arr(i, j, k) = (lx_arr(i, j, k) == 0)? 0 : 1;
-#endif
-            }
-        );
-
-    }
-}
-
-void
 WarpX::MarkExtensionCells ()
 {
     using ablastr::fields::Direction;
